Include program halt to debug with JTAG interface
We're using this because now, the board finishes it's initialization
with the MCH and it stops sending any more requests, turning the debug
much harder.
With this loop, the program stays on this line until we change this
variable value with the LPCLink Debugger interface. This way we can put
the breakpoints and follow the messages income at the speed we want.