Remove dynamic task creation from ipmi handlers
Now we just call the ipmi handler function insidethe IPMI Task.
This change was necessary because spawning a task consumes a lot of RAM
and since the messages come in a burst at the board insertion, we were
running out of heap because the garbage collector inside the Idle task
was not able to run.