Remove erase call to Flash
We don't need to erase the Flash memory before programming, since we'll
just overwrite the data present in it. Also, the HPM image is filled
with  padding bytes to prevent previous data to affect the new firmware.

PS: Bulk Erase is a very slow command, takes about 170s to complete (max 250s)