Create different linker scripts for application and bootloader binaries
Now the bootloder region starts at 0x00 and goes up to 0x2000 (8kB).
The application code begins at 0x2000 and ends at 0x10000 (56kB).
There's also a spare region in the last 2 sectors from 0x10000 to
0x11FFF reserved for firmware upgrade.
The last word of the application areas (upgrade and active) is the
firmware id.