Run dfu-util only when programming
Fixes #20
Dfu-util only returns success on its first run, if the programmer has
already been booted, the dfu-util won't be able to run successfully.
In order to not stop the compilation chain because of dfu-util 'false'
error, we ignore its output appeding '|| true' to its call, so it always
returns true.