Reformat AFC's Payload state machine
We were having a lot of problem regarding this state machine, like the
board not quiescing, turning the DCDC's on and off in a short time, etc.

Some of the state were merged and a few more control flags checking were
added on each state transition.
The full state machine will be documented later on