Change sensors input/output parameters
Now we don't return the value read from the sensor directly. Instead we
take a pointer to a variable where we're gonna save the read and return
true if the chip access was successful.
This way we avoid return a false '0' read when we can't access the chip
for some reason