[modules/eeprom_24xx64] Fix tx buffer count for segmented transfers
Since we have a limited 32-byte transaction with this EEPROM, we have to
write the page address + 32 data bytes. The way it's implemented creates
a 34-byte long buffer to hold all the bytes that need to be transfered
(including the page address).
We have to discount these 2 page address from the data buffer count.