[rtm-8sfp/rtm_user.c] Add timeout on RTM presence ping
First, default the function reading to ABSENT, since a I2C failure also
means that the board is not connected.
By adding a timeout on the bus config function we avoid reporting the
board as absent if the I2C bus is busy. 100ms should be enough time for
the firmware to free the bus.