[module/rtm.c] Perform the first hotswap reading right after detecting the board
This is needed only to maintain the LED configuration integrity in case
the MMC reboots after the MCH configured all the LEDs.
On the AMC LEDs, this is done by the hotswap task upon firmware
initialization, but since we can't configure the RTM LEDs before
detecting and initializing its hardware, we must do it here.