Add support to openocd, remove the LPCXpresso dependency
Moving to openocd simplifies the flash / debug logic by providing an
agnostic interface to different debug probes. Also it is free
software, so we avoid proprietary dependencies.

The only caveat is if you are using LPCLink / LPCLink2 debug probes
make sure it is pre-loaded with the cmsis-dap or jlink firmware.