Implement software reset independently from watchdog
When debugging via SWD/JTAG the watchdog should be disabled to avoid
resetting the mcu when it is stopped. The hpm module required a
delayed software reset that was previously done by not feeding the
watchdog. Now it is handled by a timer callback that requests a system
reset via the SCB AIRCR register (arm cortex-m specific).