[ucontroller] New bootloader for LPC17xx microcontrollers (newboot)
The original bootloader (boot/boot.c) doesn't support self-update and
is hardcoded specifically for the LPC1764 flash memory layout. newboot
runs the update routines in RAM, allowing self-update and it gets the
flash memory layout from symbols defined in the linker script, so it
is easier to use it in other LPC17xx devices. It also erases the flash
update region after a successful update, as this operation can take a
long time when done from the openMMC side and it blocks servicing ISRs
leading to IPMI failures.

newboot isn't compatible with the original HPM implementation
(port/ucontroller/nxp/lpc17xx/lpc17_hpm.c), so it needs to be updated
before replacing the older bootloader.