[newboot] Move .text section to RAM
Instead of marking every function used during the firmware update
process with an attribute to put it in a different section that is
allocated in RAM, move the entire .text section to RAM. The bootloader
can not use more than 8K of flash, so 16K of RAM is plenty for .data
and .text sections.

This also fixes problems with gcc wanting to call functions
implicitly in the functions marked with ".ramtext", breaking the
assumption that these functions are safe to call when self-updating.