Dummy handler to receive the I2CSlaveAddr when don't match the IPMBAddr
When repeatedly upgrading and downgrading the openMMC via ipmitool, it
crashes sporadically at Chip_I2C_SlaveStateHandler. This happens because
for some motive, when calling startMasterXfer() sometimes an interrupt
occurs for the I2C0 peripheral, and the Chip_I2C_SlaveStateHandler is
called, dispatching  the handling to the callback registered for
I2C_SLAVE_GENERAL. But, this callback is not configured, resulting in a
invalid memory access.

It is not clear why this interrupt is generated for an I2C slave address
other than the expected IPMB address obtained by get_ipmb_addr(), but
this should be properly handled and not lead to a crash.

Fixes #160.