Enable voltage and current sensors for AFCv4
Previously, the voltage and current sensors were disabled because it
was causing the MCH to crash with a lot of 'AMC +12V Curr' assertion /
deassertion messages and IPMB errors. The root cause for this
behavior is not yet understood, but it is somewhat related to the
voltage sensors SDR struct not been populated with the thresholds
(they were all commented). I've updated all thresholds (voltages and
currents) to reasonable defaults, but they might be wrong, so a more
in depth study of the AFCv4 power rails should be conducted to obtain
the correct values.

It works for now, and the MCH doesn't crash, neither complain about
out of range voltages or currents.