Fix RTM disable payload power error check
Commit 331bb6fea7173a9abdd14f22d4a834b3253bc334 replaced the
rtm_quiesce() function with rtm_disable_payload_power(), but the later
returns the mmc_err enum type, so it should be compared with the enum
values.

MMC_OK is the first value defined in the mmc_err enum type, so it was
represented numerically as 0, and would evaluate as false in the if
statement.