Workaround FMC XM105 hardware bug
The FMC XM105 card has a design bug due the Si570 been connected to
the MMC I2C bus, but powered from the FMC 3V3 payload power supply
instead of the 3V3 management power. This causes the Si570 to be
unpowered while the FMC payload power is down, and due the I/O
clamping diodes, it forces the SDA and SCL lines to low, causing the
I2C bus to hang forever is this state. To reconfigure the I2C mux
PCA9557, I2C communication is necessary (in-band) but if it was left
in the MUX0 selection, it can not be reconfigured.

The sensor tasks had higher priority than the payload management task,
and the way the I2C HAL is implemented, it uses a busy wait loop
(without any calls to vTaskDelay()). This causes the sensor tasks to
never give control back to lower priority tasks, but there is no way
to power the FMC 3V3 payload power supply if the vTaskPayload is never
executed.

To workaround this problem, make vTaskPayload have the same priority
of the sensors tasks.