Check payload power voltage twice before updating pp_good [afcv4.0]
We observed that a specific AFCv4.0 card suffers a loss of power in
the payload power rails event though the MCH indicates that the card
is in M4 state. It seems that the MMC detects a voltage dip in the 12V
power rail and turns off the DC/DC converters, but we suspect it is a
false reading.

This is a workaround to try avoiding turning off the payload power if
the AMC +12V voltage measurement dips below 8V for a single
measurement.